The last changes of the original BeaverDam have been deleted. This repository uses a previous version (December 2017).
Support for scene annotations has been added. These annotations are
global annotations and can be used for labeling the scene
The visibility of the scene annotations can be changed by the user.
The scene annotations have a custom value. This value is decided by the
user.
The default annotations can have several values.